refactor: 各 Fragment はパッケージで細分化する必要がなくなったため止めた。
