+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/repositorySearchFragment"
-        android:name="jp.co.yumemi.android.codecheck.ui.search.RepositorySearchFragment"
+        android:name="jp.co.yumemi.android.codecheck.RepositorySearchFragment"
         android:label="@string/app_name">
         <action
             android:id="@+id/action_repositorySearchFragment_to_repositoryDetailFragment"
@@ -15,7 +15,7 @@
 
     <fragment
         android:id="@+id/repositoryDetailFragment"
-        android:name="jp.co.yumemi.android.codecheck.ui.detail.RepositoryDetailFragment"
+        android:name="jp.co.yumemi.android.codecheck.RepositoryDetailFragment"
         android:label="@string/app_name">
         <argument
             android:name="item"
